handler: refactor memory layout & write simple allocator
What I ended up going with for now is an arena allocator for request
metadata like headers etc, and just using `malloc` for the actual
request read buffer, as obviously that needs to be continuous &
re-allocatable over time, so doesn't immediately seem to benefit from a
more specific allocator.  It's all the little random allocations over
time that we need to track anyway.